Move EXIT_CODEs to error.py and correct default error code from 1 -> 2
