+++ semgrep/semgrep/constants.py
@@ -9,7 +9,7 @@ RULES_KEY = "rules"
 SEMGREP_USER_AGENT = "semgrep-give-me-yaml"
 ID_KEY = "id"
 SEMGREP_URL = "https://semgrep.dev/"
-PLEASE_FILE_ISSUE_TEXT = "An error occurred while invoking the semgrep engine; please help us fix this by filing an an issue at https://semgrep.dev"
+PLEASE_FILE_ISSUE_TEXT = "An error occurred while invoking the semgrep engine; please help us fix this by creating an issue at https://semgrep.dev"
 
 DEFAULT_SEMGREP_CONFIG_NAME = "semgrep"
 DEFAULT_CONFIG_FILE = f".{DEFAULT_SEMGREP_CONFIG_NAME}.yml"
