+++ semgrep/semgrep/error.py
@@ -1,3 +1,13 @@
+OK_EXIT_CODE = 0
+FINDINGS_EXIT_CODE = 1
+FATAL_EXIT_CODE = 2
+INVALID_CODE_EXIT_CODE = 3
+INVALID_PATTERN_EXIT_CODE = 4
+UNPARSEABLE_YAML_EXIT_CODE = 5
+NEED_ARBITRARY_CODE_EXEC_EXIT_CODE = 6
+MISSING_CONFIG_EXIT_CODE = 7
+
+
 class SemgrepError(Exception):
     """
     Parent class of all exceptions we anticipate in Semgrep commands
@@ -6,7 +16,7 @@ class SemgrepError(Exception):
     are displayed to the user.
     """
 
-    def __init__(self, *args: object, code: int = 1) -> None:
+    def __init__(self, *args: object, code: int = FATAL_EXIT_CODE) -> None:
         self.code = code
 
         super().__init__(*args)
