+++ semgrep/semgrep/util.py
@@ -12,16 +12,6 @@ global QUIET
 DEBUG = False
 QUIET = False
 
-# Exit codes
-OK_EXIT_CODE = 0
-FINDINGS_EXIT_CODE = 1
-FATAL_EXIT_CODE = 2
-INVALID_CODE_EXIT_CODE = 3
-INVALID_PATTERN_EXIT_CODE = 4
-UNPARSEABLE_YAML_EXIT_CODE = 5
-NEED_ARBITRARY_CODE_EXEC_EXIT_CODE = 6
-MISSING_CONFIG_EXIT_CODE = 7
-
 
 def is_url(url: str) -> bool:
     try:
