+++ semgrep/tests/e2e/snapshots/test_rule_parser/test_rule_parser__failure__error_messages/badpattern/error.txt
@@ -1,1 +1,3 @@
-invalid pattern "$X == $X 3": Parsing.Parse_error
+non-zero return code while invoking semgrep-core:
+	invalid pattern "$X == $X 3": Parsing.Parse_error
+An error occurred while invoking the semgrep engine; please help us fix this by creating an issue at https://semgrep.dev
